Move AssayWheel to own file
